requirements.txt
# Core Data Processing
pandas>=2.1.4
numpy>=1.26.2

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9

# API & Web
requests>=2.31.0
beautifulsoup4>=4.12.2

# Machine Learning
scikit-learn>=1.3.2
joblib>=1.3.2

# Visualization
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.18.0

# Dashboard
streamlit>=1.29.0

# Configuration
pyyaml>=6.0.1

# Development & Testing
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.12.0
flake8>=6.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.1